feat(PubSub): Modify application for XDP
 - Modify Publisher and Loopback application for XDP
